Refactoring
- Change the path structure of the common and constant files
- Fix index.ts
