+++ backend/src/common/constant/URL.ts
@@ -1,0 +1,11 @@
+const PREFIX = {
+  API: '/api',
+};
+
+const ENDPOINT = {
+};
+
+export {
+  PREFIX,
+  ENDPOINT
+}
